requirements.txt
absl-py==2.1.0
accelerate==1.6.0
aenum==3.1.15
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-cors==0.7.0
aiosignal==1.3.2
airportsdata==20241001
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
anyio==4.9.0
astor==0.8.1
attrs==25.1.0
av==14.1.0
bitsandbytes==0.45.2
black==25.1.0
blake3==1.0.4
blis==0.7.11
cachetools==5.5.1
catalogue==2.0.10
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
cloudpathlib==0.16.0
cloudpickle==3.1.1
colorama==0.4.6
colorful==0.5.6
colorlog==6.9.0
compressed-tensors==0.9.3
confection==0.1.5
contourpy==1.3.2
cupy-cuda12x==13.4.1
cycler==0.12.1
cymem==2.0.11
DataProperty==1.1.0
datasets==3.3.1
deepspeed==0.15.4
Deprecated==1.2.18
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
einops==0.8.1
email_validator==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
fastrlock==0.8.3
filelock==3.18.0
flake8==7.1.2
fonttools==4.57.0
frozenlist==1.5.0
# fsspec==2025.5.0
ftfy==6.3.1
gguf==0.16.3
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-auth==2.38.0
googleapis-common-protos==1.67.0
grpcio==1.70.0
h11==0.14.0
hf-xet==1.1.2
hf_transfer==0.1.9
hjson==3.1.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.4
idna==3.10
importlib_metadata==8.0.0
iniconfig==2.0.0
inquirerpy==0.3.4
interegular==0.3.3
isort==6.0.0
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lark==1.2.2
latex2sympy2_extended==1.0.9
Levenshtein==0.27.1
liger_kernel==0.5.2
# lighteval @ git+https://github.com/huggingface/lighteval.git@4f381b352c0e467b5870a97d41cb66b487a2c503
llguidance==0.7.22
llvmlite==0.44.0
lm-format-enforcer==0.10.11
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
math-verify==0.6.0
matplotlib==3.10.1
mbstrdecoder==1.1.4
mccabe==0.7.0
mdurl==0.1.2
mistral_common==1.5.5
mpi4py==4.0.3
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
open_clip_torch==2.32.0
openai==1.73.0
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.11.0.86
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-semantic-conventions-ai==0.4.9
outlines==0.1.11
outlines_core==0.1.26
packaging==25.0
pandas==2.2.3
parameterized==0.9.0
partial-json-parser==0.2.1.1.post5
path==17.1.0
pathspec==0.12.1
pathvalidate==3.2.3
peft==0.15.1
pfzy==0.3.4
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
portalocker==3.1.1
preshed==3.0.9
prometheus-fastapi-instrumentator==7.0.2
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==4.25.7
psutil==7.0.0
py-cpuinfo==9.0.0
py-spy==0.4.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pycodestyle==2.12.1
pycountry==24.6.1
pydantic==2.11.3
pydantic_core==2.33.1
pyflakes==3.2.0
Pygments==2.19.1
pyparsing==3.2.3
pytablewriter==1.2.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.1
qwen-vl-utils==0.0.10
# -e git+https://github.com/Jike338/R1-V.git@abe4f03caa39b0f10a5c84217b80cf86a1fddbff#egg=r1_v&subdirectory=src/r1-v
RapidFuzz==3.13.0
ray==2.46.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.14.6
rouge_score==0.1.2
rpds-py==0.22.3
rsa==4.9
sacrebleu==2.5.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
sentry-sdk==2.22.0
setproctitle==1.3.4
shellingham==1.5.4
six==1.17.0
smart-open==6.4.0
smmap==5.0.2
sniffio==1.3.1
# spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.45.3
sympy==1.13.1
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tensorboardX==2.6.2.2
termcolor==2.3.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.9.0
timm==1.0.15
tokenizers==0.21.1
torch==2.6.0
torchaudio==2.6.0
TorchCodec==0.1.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.52.2
triton==3.2.0
trl==0.14.0
typepy==1.3.4
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.29.2
vllm==0.8.5.post1
wandb==0.18.3
wasabi==1.1.3
watchfiles==1.0.4
wcwidth==0.2.13
# weasel==0.3.4
websockets==15.0
word2number==1.1
wrapt==1.17.2
xformers==0.0.29.post2
xgrammar==0.1.18
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
flash_attn==2.7.4.post1